Implement parser for Consultation, Vaccination, and Surgery type appointments.
