--- src/main/java/functionalities/Storage.java
@@ -1,3 +1,0 @@
-package functionalities;
-public class Storage {
-}
