+++ src/main/java/functionalities/appointments/Vaccination.java
@@ -1,0 +1,28 @@
+package functionalities.appointments;
+
+import functionalities.Animal;
+import functionalities.Owner;
+
+public class Vaccination extends Appointment {
+    protected String date;
+    protected String time;
+    protected String vaccine;
+
+    public Vaccination(String uid, Animal animal, Owner owner, String date, String time, String vaccine) {
+        super(uid, animal, owner);
+        this.uid = uid;
+        this.animal = animal;
+        this.owner = owner;
+        this.date = date;
+        this.time = time;
+        this.vaccine = vaccine;
+    }
+
+    @Override
+    public String toString() {
+        return " UID: " + uid + " | vaccine: " + vaccine + '\n'
+                + " Date: " + date + " | Time: " + time + '\n'
+                + " Animal Name: " + animal.toString() + '\n'
+                + " Owner Name: " + owner.toString();
+    }
+}
