+++ src/main/java/functionalities/commands/ExitCommand.java
@@ -2,6 +2,7 @@ package functionalities.commands;
 
 import exception.SniffException;
 import functionalities.SniffTasks;
+import functionalities.ui.Ui;
 
 public class ExitCommand extends Command {
 
@@ -11,5 +12,6 @@ public class ExitCommand extends Command {
     @Override
     public void executeCommand(SniffTasks tasks) throws SniffException {
         this.isExit = true;
+        Ui.showUserMessage(" Bye, hope to see you again soon!");
     }
 }
