+++ src/main/java/functionalities/commands/RemoveCommand.java
@@ -2,9 +2,10 @@ package functionalities.commands;
 
 import exception.SniffException;
 import functionalities.SniffTasks;
-import functionalities.Ui;
+import functionalities.ui.Ui;
 
 public class RemoveCommand extends Command {
+
     public String uid;
 
     public RemoveCommand(String uid) {
@@ -14,6 +15,6 @@ public class RemoveCommand extends Command {
     @Override
     public void executeCommand(SniffTasks tasks) throws SniffException {
         tasks.removeAppointment(uid);
-        Ui.showUserMessage(" Task removed successfully");
+        Ui.showUserMessage(" Task removed successfully!");
     }
 }
