+++ src/main/java/functionalities/commands/SurgeryCommand.java
@@ -4,10 +4,10 @@ import exception.SniffException;
 import functionalities.Animal;
 import functionalities.Owner;
 import functionalities.SniffTasks;
-import functionalities.Ui;
+import functionalities.ui.Ui;
 
 public class SurgeryCommand extends Command {
-    private final String uid;
+
     private final Animal animal;
     private final Owner owner;
     private final String priority;
@@ -16,11 +16,9 @@ public class SurgeryCommand extends Command {
     private final String endDate;
     private final String endTime;
 
-    public SurgeryCommand(String uid, String animalType, String animalName,
-                          String ownerName, String contactNumber, String startDate,
-                          String startTime, String endDate, String endTime,
-                          String priority) {
-        this.uid = uid;
+    public SurgeryCommand(String animalType, String animalName, String ownerName,
+                          String contactNumber, String startDate, String startTime,
+                          String endDate, String endTime, String priority) {
         this.priority = priority;
         this.animal = new Animal(animalType, animalName);
         this.owner = new Owner(ownerName, contactNumber);
@@ -33,6 +31,6 @@ public class SurgeryCommand extends Command {
     @Override
     public void executeCommand(SniffTasks tasks) throws SniffException {
         tasks.addSurgery(animal, owner, priority, startDate, startTime, endDate, endTime);
-        Ui.showUserMessage(" Surgery added successfully");
+        Ui.showUserMessage(" Surgery added successfully!");
     }
 }
