+++ src/main/java/functionalities/commands/VaccinationCommand.java
@@ -4,20 +4,17 @@ import exception.SniffException;
 import functionalities.Animal;
 import functionalities.Owner;
 import functionalities.SniffTasks;
-import functionalities.Ui;
+import functionalities.ui.Ui;
 
 public class VaccinationCommand extends Command {
-    private final String uid;
     private final Animal animal;
     private final Owner owner;
     private final String vaccine;
     private final String date;
     private final String time;
 
-    public VaccinationCommand(String uid, String animalType, String animalName,
-                              String ownerName, String contactNumber, String vaccine,
-                              String date, String time) {
-        this.uid = uid;
+    public VaccinationCommand(String animalType, String animalName, String ownerName,
+                              String contactNumber, String vaccine, String date, String time) {
         this.animal = new Animal(animalType, animalName);
         this.owner = new Owner(ownerName, contactNumber);
         this.vaccine = vaccine;
@@ -27,7 +24,7 @@ public class VaccinationCommand extends Command {
 
     @Override
     public void executeCommand(SniffTasks tasks) throws SniffException {
-        tasks.addConsultation(animal, owner, date, time);
-        Ui.showUserMessage(" Surgery added successfully");
+        tasks.addVaccination(animal, owner, date, time, vaccine);
+        Ui.showUserMessage(" Vaccination added successfully!");
     }
 }
