+++ src/main/java/functionalities/storage/Storage.java
@@ -1,0 +1,3 @@
+package functionalities.storage;
+public class Storage {
+}
